// run-archive/scenario-2/1777107333568-scenario-2-deepseek-v4-pro-high-plan-medium-build-attempt-1/frontend/src/api/model/bookingRequest.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * RestaurantBooking.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface BookingRequest {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  restaurantId: number | string;
  customerName: string;
  customerEmail: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  partySize: number | string;
  date: string;
  startTime: string;
}